requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.1.3
django-cpf-cnpj==1.0.0
django-crontab==0.7.1
django-debug-toolbar==3.2.4
django-google-tag-manager==0.0.5
django-loginas==0.3.10
django-session-security==2.6.6
idna==3.3
newrelic==5.18.0.148
psycopg2-binary==2.9.3
python-decouple==3.5
pytz==2021.3
requests==2.27.1
sentry-sdk==1.5.2
sqlparse==0.4.2
stripe==2.64.0
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
gunicorn==20.0.4